错误代码修改

+++ masterdata-core/src/main/java/com/accenture/masterdata/core/outEntity/OrganizationOut.java
@@ -12,26 +12,26 @@ import lombok.Data;
 @Data
 public class OrganizationOut {
 
-	private int id;
+	private Long id;
 	private String name;
 	private String code;
 	private String comment;
-	private int parentId;
+	private Long parentId;
 	private String parentName;
 	private String hierarchyName;
-	private int hierarchyLevel;
-	private int hierarchyId;
-	private int creatorUserId;
+	private Long hierarchyLevel;
+	private Long hierarchyId;
+	private Long creatorUserId;
 	private String likeCode;
 	private String creatorName;
 	private Date creationTime;
-	private int lastModifierUserId;
+	private Long lastModifierUserId;
 	private Date lastModificationTime;
 	private boolean isDeleted;
-	private int deleterUserId;
+	private Long deleterUserId;
 	private String deleterName;
 	private Date deletionTime;
-	private int tenantId;
+	private Long tenantId;
 	private String tenantName;
 	private List<OrganizationOut> child;
